Don't clobber the selection in insertTag()
I've thought better of the earlier behavior.  Especially since the
action of insertTag blows away the undo buffer, having the selection
deleted is unlikely to be desirable.

The new behavior will insert the tag at the start of the selection, but
leave the cursor at the end of the selection.  The action of
insertXMLTag and insertWrapperTag should not be affected.