Fix click-and-play requiring a page refresh each time
When navigating segments, the page retains the old WAVE element, so the
even handler still triggers.  However, the page apparently deletes and
recreates the SPAN element serving as the play button.  So, keeping the
one that existed at handler adding time does not work.  This patch
causes the click handler to fetch the SPAN again each time.